feat(async): clean __init__.py
