+++ .gitignore
@@ -9,3 +9,4 @@ docs/_build
 .testrepository/
 .tox
 venv/
+tags
