+++ gitlab/client.py
@@ -1,9 +1,43 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2013-2017 Gauvain Pocentek <gauvain@pocentek.net>
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Lesser General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""Wrapper for the GitLab API."""
+
+import importlib
 from typing import Any
 
+import gitlab.config
 import httpx
 from gitlab import exceptions as exc
+from gitlab import utils
 from gitlab.exceptions import on_http_error
 from gitlab.types import GitlabList
+
+REDIRECT_MSG = (
+    "python-gitlab detected an http to https redirection. You "
+    "must update your GitLab URL to use https:// to avoid issues."
+)
+
+
+def _sanitize(value):
+    if isinstance(value, dict):
+        return dict((k, _sanitize(v)) for k, v in value.items())
+    if isinstance(value, str):
+        return value.replace("/", "%2F")
+    return value
 
 
 class BaseGitlab:
